Revert "Fix Manually-Enterd First Sponsor Selection on Review (#21190)"
This reverts commit 34ac3289fb7c077ca11b931f60db8e97fecd2f40.
